Rewrite verify_findMD5Sum
The wii was crashing in the call to `mxmlIndexNew`. Looking at the source code to this function, it does a realloc every 64 elements. I'm guessing all the reallocs is causing the crash. Maybe the wii just stops giving us memory? Anyways, to avoid the issue, I rewrote the search.

We were doing a linear search to build an index so we can do a linear search. It should be faster to do a linear search for the md5 value directly rather than building the index.

The rewrite then pulls the name from the parent of the node with the md5 value hit. This change should be faster and non-heap allocating.

I haven't tested the new code on real hardware yet. I've just been testing mxml code in a sample app on my desktop.