Generate md5 and sha1 values lowercase
The lookup into the xml is now case sensitive.  As far as I can tell, the dat files we support always have md5 and sha1 values that are lowercase so a case insensitive search isn't "needed" so long as our input is lowercase. WIthout this fix, the verify_findMD5Sum call fails to find the value because the md5 values differ by casing.